Changes ...
Found bug where we forgot to filter the values of the Root DSE entry.  This
was causing several clients to fail when accessing the Root DSE for specific
entries.

If we find that we need to do this regularly such as with subschemaSubentries
then we might as well centralize the code for filtering singleton entries 
before returning them.



git-svn-id: https://svn.apache.org/repos/asf/incubator/directory/eve/trunk@109847 13f79535-47bb-0310-9956-ffa450edef68