While fixing the problem with the broken server.disable.anonymous property I noticed that anonymous binds are still possible with the bind operation.  Here's why ...
In LDAP an anonymous bind is when an operation is requested without having had
a successful bind take place first.  For example a search can be attempted 
before bothering with performing a bind op.  This feature was there for all reqs
minus the bind request.  So you could bind first anonymously to establish a 
logical session then execute a command like search as the user "" with no 
credentials.  This was a security hole. 

I fixed this by preventing a bind as well as other operations when certain 
conditions indicative of an anonymous bind are in effect.  


git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@158946 13f79535-47bb-0310-9956-ffa450edef68