changes ...
 o moved server context factory ICF impl to main project to relieve deps on core
 o removed need for initial load of LDIF file altogether
 o moved SCF tests to main

Note: there really is not much now to SCF.  It's there out of convenience now
and most of what it does can be handled by a main.  However we're leaving it 
in place out of convenience for those that want to be able to start up the
server using JNDI and have all the wire protocols included.



git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@165308 13f79535-47bb-0310-9956-ffa450edef68