Added some matching rules that were in the extra matchingRules RFC here:
   http://www.faqs.org/rfcs/rfc3698.html

This is just too backwards to do we have to fix the schema subsystem to make
it work better for extention.  It was never finished and this is why we have
the runnaround. 

This was a fix for daniel.armbrust.list@gmail.com.  Next time we need a JIRA
for it too.



git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@171150 13f79535-47bb-0310-9956-ffa450edef68