Added some schema checks to prevent improper modify operations from succeeding. Namely we added checks to satisfy the modify tests submitted by Stefan in the following JIRA DIREVE-232 issue:
http://issues.apache.org/jira/browse/DIREVE-232

His test case was added as well as the changes to make sure the tests pass. 

NOTE: most of these fixes testing for schema correctness should be moved into
the schema checking interceptor however there may be some information loss 
without access to the request - this should be inspected before moving this
code into the interceptor.



git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@264063 13f79535-47bb-0310-9956-ffa450edef68