overview ...
Added a new feature to the interceptor mechanism which allows interceptors to be 
bypassed when calls are made to the nexus proxy.  This feature allows interceptors
to issue operations against the proxy with the names of interceptors to be 
bypassed.  A wildcard is also included which bypasses all interceptors in the
chain.  

This feature is exposed to interceptors through alterations to the Invocation
class which now exposes the nexus proxy as well as other information about the
call.  It is intended to be used to prevent possible infinite recursion when
interceptors access the DIT.  It may also be used to allow privileged access
as well as to prevent incurring unnecessary inefficiencies.

changes ...

 o merging back changes made on the "interceptor-bypass" branch onto
   the trunk
 o added the nexus proxy to the Invocation object
 o added a collection of Interceptor names to Invocation to bypass
 o modified InterceptorChain to skip Interceptors when determining 
   the next interceptor and when determining the starting interceptor
 o overloaded operations on the proxy to also take a collection of
   interceptor names to bypass 



git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@326026 13f79535-47bb-0310-9956-ffa450edef68