Fixed bug arrising from corrections made to fix DIRSERVER-1019 ...
Seems as though the fixes for this revealed an inconsistent state that occurs in
the poorly written filter parser.  I have created another JIRA to rewrite the 
filter parser here:

   https://issues.apache.org/jira/browse/DIRSERVER-1020

Anyways I found a work around which hides the symptoms instead of curring this 
disease.  The cure is a parser rewrite in the nice way that seelmann did it for
the other parsers.  The work around I used was to make the filter parser in the
ServerDirContext implementation a instance member instead of a static class 
member.  Again this just hides the issue rather than fixing it.



git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@566769 13f79535-47bb-0310-9956-ffa450edef68