Adding new project module to contain tools used for dealing with xdbm based partition stores.  XDBM means and partition based on two column datastructures where keys are sorted.  XDBM partition and store implementations use a standard scheme for a partition of the Directory Information Base (DIB).  This scheme is documented in the developer documentation for the JDBM based partition impl. This scheme is the one which has a master table, with system (required) and user indices into the master table.
So this module is just a place to store utilities like the gui used to inspect 
such store implementations, and their indices.  These tools operate on xdbm 
constructs like Index, Table and MasterTable as opposed to implementation 
specific constructs.  This is intentional so the utilities and tools can be 
used across implementations.  Some other implementations may be based on AvlTree
(in memory) or other btree implementations on disk.


git-svn-id: https://svn.apache.org/repos/asf/directory/sandbox/akarasulu/bigbang/apacheds@640528 13f79535-47bb-0310-9956-ffa450edef68