changes to merge reads from master and index reads with txn log:
logedits: log edits to keep track entry modifications, entry add/delete and index changes. 

ReadWriteTxn and TxnLogManager: As data log edits are added a summary of index changes for the current txn is built. This summary is of the for<partitionDN, attributeOID, sorted set of index changes>

TxnIndexCursor: this implements a cursor interface for a single txn for a single index. Since index changes are kept memory, this cursor is built using an iterator.

IndexCursorWrapper.java: This wraps a partition index and creates a transactionally consistent view of it using the wrapped index and TxnIdexCursors for the txns that the current txn depends on.This basically implements a merge sort of various index cursors.

TODO: finish MastertableWrapper, add IndexWrapper


git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/branches/apacheds-txns@1185638 13f79535-47bb-0310-9956-ffa450edef68