added IndexWrapper and completed MasterTableWrapper.
For index cursors,  null values are used skip whole set of values corresponding to a key. Also an index cursor can be created to traverse just over the values corresponding to a key. I changed TxnIndexCursor to account for these cases.

IndexWrapper mostly uses IndexCursorWrapper to do lookups and return cursors.

Master table wrapper used mergeUpdates of the log manager to get the current view of the entry it reads from the wrapped partition.



git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/branches/apacheds-txns@1186778 13f79535-47bb-0310-9956-ffa450edef68