Changes to add conflict detection service to the txn manager.
DnSet: keeps track of the base dn and the scope a read depends on or a write afffects.

For a txn, readset is always a superset of writeset.

When txn tries to commit, its read set is checked against the writeset of the txns that committed after it started. If conflict is found, a txnconflict exception is thrown.



git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/branches/apacheds-txns@1189768 13f79535-47bb-0310-9956-ffa450edef68