Fix a failing test with the latest JVM version on mac (1.6.0_29).
This is absolutely ugly : the name of a private field for the sun.security.krb5.KrbKdcReq class has changed from udpPrefLimit to defaultUdpPrefLimit, leading to a NoSuchFieldException when we try to fetch this field to change its value. I don't know why we try to modify this value, but in any case, the resulting failure is just expected.

I don't know if it's possible to avoid such usage of reflection o get the same result, but if so, we must absolutely avoid accessing inner fields of classes we don't own ...


git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@1199700 13f79535-47bb-0310-9956-ffa450edef68