Added DefaultOperatinExecutionManager.java: Executes the modification and lookup operations in a transactionally consistent manner. For modification operatiopns, prepares the txn log edits.
Handling of txn log edits for modify(add, remove, replace atrributes), add, delete is complete. Next set of changes will be for move and rename.

This execution manager can work with any partition as long as the partition exposes the master table and the required system indices.

Other changes are mostly to support this new class.




git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/branches/apacheds-txns@1201834 13f79535-47bb-0310-9956-ffa450edef68