changed jdbm partition test to use the partition interface and operation execution manager when doing queries and modifiications.
A problem I detected was the possible inconsistency when reading btree redirects from anohter btree. Say we have Btree A and one of the records of this btree points to Btree B. We open a browser on tree A as of time 10. Then at some time we read a pointer to Btree B. From BTree A, we get this pointer as of time 10 and then use it to read Btree B say at time 15.. However, the new search for BTree B will as of time 15 and by that time that pointer might have been assigned to some other object.  This is unlikely but the possibility is there. To work around this problem we avoid deleting the Btree pointer from the underlying record manager even when Btree B is deleted. This is just a work around and there should be a better way to solve this problem. 


git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/branches/apacheds-txns@1208333 13f79535-47bb-0310-9956-ffa450edef68