Various fixes based on the tests done for txns.
Also added methods to merge index lookups and checks for existince of an entry in a more efficient way. Used to construct a cursor each time a lookup is done. Thought this might be too much of a overhead, so added methods to directly merge what is read from partitions with the changes in the txn log.


git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/branches/apacheds-txns@1208350 13f79535-47bb-0310-9956-ffa450edef68