It seems the OSGi code that was merged into the transaction branch was depending on the newer M7 ApacheDS code base while the Txn branch is based on and older M4 release. This might explain why there might have been some issues with tests.
We have to back port the OSGi code to the M4 code base to get it working with
the Txn code. This just sets the projects straight but also introduces some 
compilation problems with the StoreConfigManager in the OSGi hub core project.

I think it might be a good idea to use a non-M?-SNAPSHOT designator to prevent
weirdness with trunk or other branches. I began updating this until I ran into
the compilation problem in the hub core. I thought I did something wrong, then
later I realized that it had not been backported to M4.   



git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/branches/apacheds-txns@1368883 13f79535-47bb-0310-9956-ffa450edef68